ipc fixes plus more others

+++ clientserver/core/common.hpp
@@ -50,7 +50,7 @@ enum Consts{
 	MAXTIMEOUT = (0xffffffff>>1)/1000
 };
 
-typedef std::pair<uint32, uint32> ObjectUidTp;
+typedef std::pair<ulong, uint32> ObjectUidTp;
 typedef std::pair<uint32, uint32> CommandUidTp;
 typedef std::pair<uint32, uint32> FileUidTp;
 typedef std::pair<uint32, uint32> RequestUidTp;
